requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.8
python-dotenv>=1.0.0
openai>=1.10.0,<2.0.0
requests>=2.25.1